Refactor geoshapes service structure
1) Split geoshapes.js into index.js, lib/geoshape.js, and queries.yaml
2) Separate Geoshapes logic from service configuration
3) Load queries through .yaml file to make code cleaner
4) Brake index.js into several function to explicitly call step by step
and improve code clarity
5) Clean unnecessary code
