+++ package.json
@@ -2,7 +2,7 @@
   "name": "@kartotherian/geoshapes",
   "version": "1.0.3",
   "description": "Kartotherian service to generate geometric shapes from PostgreSQL data",
-  "main": "geoshapes.js",
+  "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -18,12 +18,13 @@
     "access": "public"
   },
   "dependencies": {
+    "@kartotherian/err": "^0.0.3",
     "bluebird": "^3.5.0",
-    "pg.js": "^4.1.1",
+    "js-yaml": "^3.12.2",
     "pg-promise": "^5.6.4",
+    "pg.js": "^4.1.1",
     "preq": "^0.5.2",
     "topojson": "^2.2.0",
-    "wd-type-parser": "^0.0.3",
-    "@kartotherian/err": "^0.0.3"
+    "wd-type-parser": "^0.0.3"
   }
 }
